refactor elements done
