+++ src/components/custom/other-platforms.js
@@ -2,8 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { SectionContainer, FlexGridContainer } from 'components/containers'
-import { Card } from 'components/elements/card'
-import { Header } from 'components/elements/typography.js'
+import { Card, Header } from 'components/elements'
 import { localize, LocalizedLink } from 'components/localization'
 import device from 'themes/device'
 // icons
