+++ src/components/custom/signup.js
@@ -3,9 +3,8 @@ import styled from 'styled-components'
 import Cookies from 'js-cookie'
 import PropTypes from 'prop-types'
 import { FlexGridContainer } from 'components/containers'
-import { Header, Text } from 'components/elements/typography'
+import { Header, Text, Image } from 'components/elements'
 import { localize, Localize } from 'components/localization'
-import Image from 'components/elements/image'
 import Wrapper from 'components/containers/wrapper'
 import Button from 'components/form/button'
 import Input from 'components/form/input'
