+++ src/components/elements/index.js
@@ -1,5 +1,5 @@
 import AutoCarousel from './auto-carousel'
-import { Card, CardChildren } from './card'
+import { Card, CardChildren, CardStyle } from './card'
 import Carousel from './carousel'
 import Divider from './divider'
 import DotLoader from './dot-loader'
@@ -8,8 +8,8 @@ import Hero from './hero'
 import Image from './image'
 import Link from './link'
 import LinearLoader from './loader'
-import Modal from './modal'
-import OffCanvas from './off-canvas-menu'
+import Modal, { useModal } from './modal'
+import OffCanvasMenu, { moveOffCanvasMenu } from './off-canvas-menu'
 import SideTab from './side-tab'
 import { Text, LinkText, Header } from './typography'
 
@@ -17,6 +17,7 @@ export {
     AutoCarousel,
     Card,
     CardChildren,
+    CardStyle,
     Carousel,
     Divider,
     DotLoader,
@@ -26,7 +27,9 @@ export {
     Link,
     LinearLoader,
     Modal,
-    OffCanvas,
+    useModal,
+    OffCanvasMenu,
+    moveOffCanvasMenu,
     SideTab,
     Text,
     LinkText,
