+++ src/pages/about/_our-goals.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { SectionContainer, FlexGridContainer } from 'components/containers'
-import { Card } from 'components/elements/card.js'
+import { Card } from 'components/elements'
 import { localize } from 'components/localization'
 import device from 'themes/device'
 import VisionSVG from 'images/svg/vision.svg'
