+++ src/pages/about/index.js
@@ -5,10 +5,8 @@ import { OurNumbers } from './_our-numbers'
 import { OurGoals } from './_our-goals'
 import { localize, WithIntl } from 'components/localization'
 import Layout from 'components/layout/layout'
-import Hero from 'components/elements/hero.js'
-import { Header } from 'components/elements/typography.js'
+import { Hero, Header, Modal, useModal } from 'components/elements'
 import Button from 'components/form/button'
-import Modal, { useModal } from 'components/elements/modal'
 import SignupModal from 'components/custom/signup-modal'
 import { Show, SEO } from 'components/containers'
