+++ src/pages/check-email/_icon-grid.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import device from 'themes/device'
-import { Text } from 'components/elements/typography.js'
+import { Text } from 'components/elements'
 import { localize } from 'components/localization'
 // Icons
 import Spam from 'images/svg/spam.svg'
