+++ src/pages/check-email/index.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { IconGrid } from './_icon-grid'
 import Button from 'components/form/button'
 import Layout from 'components/layout/layout'
-import Modal, { useModal } from 'components/elements/modal'
+import { Modal, useModal } from 'components/elements'
 import SignupModal from 'components/custom/signup-modal'
 import { GridContainer, SectionContainer, SEO } from 'components/containers'
 import { localize, WithIntl } from 'components/localization'
