+++ src/pages/dbot/_tutorial.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import Image from 'components/elements/image'
 import { SectionContainer, FlexGridContainer } from 'components/containers'
-import { Header, Text } from 'components/elements/typography.js'
+import { Header, Text, Image } from 'components/elements'
 import { localize } from 'components/localization'
 import device from 'themes/device'
