+++ src/pages/dtrader/_hero.js
@@ -3,10 +3,8 @@ import styled from 'styled-components'
 import { localize } from 'components/localization'
 import { Container, Show } from 'components/containers'
 import device from 'themes/device.js'
-import { Header } from 'components/elements/typography.js'
-import Modal, { useModal } from 'components/elements/modal'
+import { Header, Image, Modal, useModal } from 'components/elements'
 import SignupModal from 'components/custom/signup-modal'
-import Image from 'components/elements/image'
 import Button from 'components/form/button.js'
 import dtrader_logo from 'images/svg/dtrader-32.svg'
