+++ src/pages/help-centre/_search-results.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { Ul, Li, ListWithLinks } from './_list'
-import { Text, Header } from 'components/elements/typography'
+import { Text, Header } from 'components/elements'
 import { localize } from 'components/localization'
 import device from 'themes/device'
