+++ src/pages/home/_how-it-works.js
@@ -7,7 +7,7 @@ import {
     FlexGridContainer,
 } from 'components/containers'
 import device from 'themes/device'
-import { Header, Text } from 'components/elements/typography.js'
+import { Header, Text } from 'components/elements'
 import { localize } from 'components/localization'
 // Icons
 import Practice from 'images/svg/practice.svg'
