+++ src/pages/home/_markets.js
@@ -2,8 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { StyledSubHeader } from './_headers'
-import SideTab from 'components/elements/side-tab'
-import { Header, Text } from 'components/elements/typography.js'
+import { Header, Text, SideTab } from 'components/elements'
 import { localize } from 'components/localization'
 import { Container, SectionContainer } from 'components/containers'
 // Icons
