+++ src/pages/home/_why-deriv-mobile.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Header } from 'components/elements/typography'
+import { Header, Modal, useModal } from 'components/elements'
 import Button from 'components/form/button'
-import Modal, { useModal } from 'components/elements/modal'
 import { localize } from 'components/localization'
 import Checklist from 'images/svg/checklist.svg'
 import SignupModal from 'components/custom/signup-modal'
