+++ src/pages/reset-password/index.js
@@ -4,7 +4,7 @@ import { Formik, Form } from 'formik'
 import Layout from 'components/layout/layout'
 import { localize, WithIntl } from 'components/localization'
 import { Container, SEO } from 'components/containers'
-import { Header, Text } from 'components/elements/typography'
+import { Header, Text } from 'components/elements'
 import Input from 'components/form/input'
 import Button from 'components/form/button'
 import validation from 'common/validation'
