+++ src/pages/terms-and-conditions/_styles.js
@@ -1,6 +1,5 @@
 import styled from 'styled-components'
-import { Header, Text } from 'components/elements/typography'
-import StyledLink from 'components/elements/link'
+import { Header, Text, Link } from 'components/elements'
 import device from 'themes/device'
 
 export const StyledSection = styled.section`
@@ -47,7 +46,7 @@ export const StyledOrder = styled.ol`
     }
 `
 
-export const TncLink = styled(StyledLink)`
+export const TncLink = styled(Link)`
     font-weight: bold;
     color: var(--color-black-3);
     font-size: var(--text-size-s);
